chore: assert on error codes, uid and type in EnrollmentAttributeValidationHookTest
Took the idea from ErrorReportContainer to add a helper into ValidationErrorReporter to ease asserting that
a particular error is present. This gives us the flexibility to assert on one or many fields. anyMatch also
keeps the test flexible in case more errors are in the error list.

The hasErrorReport/WarningReport code can be moved to TrackerValidationReport once ValidationErrorReporter is removed