chore: assert on error/warning codes, type, uid in tracker error reports TECH-880 (#9531)
so later refactoring of how errors are inserted into the ValidationErrorReporter is safe. Added a few tests for more
error cases

Took the idea from ErrorReportContainer to add a helper into ValidationErrorReporter to ease asserting that
a particular error is present. This gives us the flexibility to assert on one or many fields. anyMatch also
keeps the test flexible in case more errors are in the error list.

The hasErrorReport/WarningReport code can be moved to TrackerValidationReport once ValidationErrorReporter is removed

* chore: unit test DefaultTrackerValidationService

to ensure report generation and removing of invalid DTOs from the bundle
works after the upcoming refactoring of ValidationErrorReporter