chore: inline all one-liner methods related to the TrackerBundle
Remove as much as possible from TrackerImportValidationContext, so
we can remove the class in the end. Passing in the TrackerBundle
into hooks should be enough.

TrackerImportValidationContext is a class that serves as a utility. You
can see that hooks mostly reach for the TrackerPreheat which is a field
on the TrackerBundle. In a few cases, hooks need information from the
TrackerBundle itself. An older TODO comment already questioned the
TrackerImportValidationContext existence. After the PR is merged
TrackerImportValidationContext will still hold a cache that we might be
able to move into the TrackerPreheat.