chore: remove dependency on TrackerBundle from TrackerErrorReport
The TrackerIdScheme from the TrackerBundle is needed if you want to add
an argument of type IdentifiableObject to an error/warning. Because of
this and the structure of TrackerReportUtils every hook that wanted to
add an error was forced to provide the TrackerBundle into the
error/warning builders. This also lead to helpers in
AbstractTrackerDtoValidationHook as adding errors was cumbersome.

Moving the logic from TrackerReportUtils into the error/warning builders
* adds typed methods instead of only relying on Object and isAssignableFrom
* cleanly separates dependencies. AbstractTrackerDtoValidationHook is
  free from responsibilties of adding errors which should solely lie
  within the TrackerValidationReportTest
* makes adding errors easier, since we do not have to always provide
  the TrackerBundle
* allows us to transition away from ValidationErrorReporter in favor of
  TrackerValidationReport