fix: Make flyway script idempotent to avoid manual dev db changes (2.38) (#9668)
The specific script was accidentally first added under the 2.37 folder which was caused to already execute in the play dev db. The script was now moved to 2.38 which caused the script to execute again and now failing in dev because it is not idempotent. Changed the script to make it idempotent to overcome this situation.