chore: TrackerIdScheme parameter is unused
the only effect on behavior it could have is that we do not add the identifier
to the map of identifiers. That would only be the case if the parameter is null,
which it never is. All identifiers in the params have a default of UID.
Of course someone could set them to null, but that would cause all sorts of other
bugs that we will notice