fix: use idScheme in event dataElement checks DHIS2-12282 (#9682) (#9723)
* chore: fix IDE warnings in EventDataValuesValidationHookTest

* chore: clean EventDataValuesValidationHookTest

* remove unnecessary mockito stubbings to use mockito mode strict and keep future tests
clean

* split tests that contained multiple test cases into multiple tests

* fix: use idScheme in dataElement programStage check DHIS2-12282

* fix: use idScheme in mandatory dataElement check DHIS2-12282

* chore: map EventDataValue.dataElement to UID

At this point EventDataValuesValidationHook already checked whether the
dataElement specified by the user exists by looking in the preheat. If
not found it would add error E1304 'DataElement `{0}` is not a valid
data element' and fail validation. Looking the dataElement up in the
preheat already respects the idScheme.

Lookup the dataElement in the preheat using the user specified identifier.
Use the UID when mapping to an EventDataValue. This means any code using
the EventDataValue.identifier from then on can rely on it being a UID.
The NTI request idScheme is not available in the program rule engine so
any existing code already assumed it was a UID.

* chore: invert condition for readability

use guard clause to reduce nesting
create arraylist with Java only instead of relying on a library for it