chore: reduce indirection in DefaultTrackerBundleService (#9711)
* chore: reduce indirection in DefaultTrackerBundleService

The pattern we often use of using a map of TrackerTypes to functions
introduces indirection making the code hard to read. We also rely on a
JAVA EE feature @PostConstruct to build the map once all dependencies
have been wired.

In this case the alternative code is straightforward to read. Removing
@PostConstruct makes testing also easier: no need for Spring doing
executing the @PostConstruct or us manually calling initMaps after the
constructor.

* chore: introduce local var for the report