fix: attributeOptionCombo is not required in NTI DHIS2-12460 (#9747)
* chore: make EventCategoryOptValidationHookTest use mockito strict
* chore: make PreCheckDataRelationsValidationHookTest use mockito strict
* chore: move getters close to preheat
TrackerImportValidationContext contains a lot of getters that should live in the bundle or preheat.
They clearly reach for data that is in the bundle or preheat

* chore: unit test validation of event attribute option combo logic

* fix: ProgramMapper does not map categoryCombo

the implementation generated by mapstruct created

if ( program.hasCategoryCombo() ) {
            program1.setCategoryCombo( program.getCategoryCombo() );
        }

where hasCategoryCombo is only true for non default category combo.

Every program has a categoryCombo. Not mapping this field leads to a
loss of information. Making for example validation in tracker
more complicated

* chore: fix bugs

* fix: set the event AOC if not provided

only in the case when we can unmistakingly figure it out

* chore: return E1055 also when AOC is given

* fix: integration test testNoCategoryOptionComboSet

the event AOC was actually a CO, so the test failed for another reason
now that the error messages are more specific.

Using only a CO not in the event program CC triggers expected error

* fix: use event.AOC for default program.CC containing multiple AOCs

* chore: add another error to make case clear

* chore: clarify error message E1117