chore: name clearly Sync/AsyncImporter
as thats what distinguishes them and its what they do import.

import() would be a nice interface method name, but its a reserved
keyword. So stick with importTracker like in our services.

We are not importing a report which is what importReport suggests. We
are merely returning a report after the import is done.