chore: remove unused SupplierDependsOn
the order in which preheat suppliers are called is solely determined by TrackerPreheatConfig.preheatOrder
Since SupplierDependsOn is not used to determine order it does not even serve as a documentation. This can be
seen by looking at the suppliers, most of which depend on the ClassBasedSupplier but only 2 actually using
the annotation.