test: quiet logging of tests to console
our logs during testing show a lot of errors/stacktraces that are part
of expected behavior. They are noise when trying to look for a failing
test and its assertion errors.

Provide a test log4j2 config via the system property log4j2.configurationFile
so that our programmatic logging config Log4JLogConfigInitializer will
cease to change the config. Only then are we in full control of logging
during testing.

Provide a muted console appender so a dev can quickly get logs into the
console when trying to reproduce a failed test where logs are
beneficial.