fix: dont mutate the encryption status in H2DhisConfigurationProvider
the integration test set the encryption status without setting it back.
If it ran before the TrackedEntityProgramAttributeEncryptionTest the
password was still "missing" even though the PostgresDhisConfigurationProvider
and its h2TestConfig.conf has a password set.

Remove the integration test as the exact case is already unit tested in
TrackedEntityAttributeValidationHookTest.shouldFailEncryptionStatus

This way we can remove the possibility of mutating the H2DhisConfigurationProvider
encryption status and future errors like this.