ci: remove e2e containers and volumes
just in case. They might be causing these unexcepted errors we see
in e2e tests. Not sure what cleanup runners perform on docker.
So maybe a DB volume of an older test is still lying around.
TrackedEntityInstanceAclReadTests seems to make changes that are
affecting the MetadataImportTest we currently see failing