test: use assertNoErrors instead of assert on size and content
in case the expected and actual size differs, we do not get more context other than that they are not equal.
Using our own assertNoErrors assertion we do also get the list of errors. This context is especially helpful
in integration tests which take long to run and/or with flaky tests only failing in CI