test: add context to assertions on error and errors size (#11326)
in case the expected and actual size differs, we do not get more context
other than that they are not equal. We usually have such assertions
before the assertions checking for error codes within the list of errors
which will then not get executed. Add context using JUnit 5 assertAll
and some custom assertions so both the check for size and content are
executed.

This context is especially helpful in integration tests which take long
to run and/or with flaky tests only failing in CI.