test: add own log4j2 test config for integration tests
* integration tests were not logging due to https://www.slf4j.org/codes.html#noProviders
* added own log4j2-test.xml for integration tests to iterate more
  quickly and customize it for integration tests
* adapt log pattern so its easier to understand where the log is coming
  from

Integration tests often need debugging since they tend to be more flaky
or only fail in CI. Logging is helpful in that case. The log4j2-test.xml
config we have in the dhis-support-test module is useful so not every
module needs to bring its own. Integration tests logs might need to be
configured differently. We use different tools like testcontainers, ...

Previously, dhis-support-test module needed to be re-installed for every
change to log4j2-test.xml. This is still needed for unit tests, but not
anymore for integration tests.

Only logging a classname like `WaitingConsumer` will leave you wondering
what library is logging the message if the message does not reveal it
itself. Added the package to the log entry. Used a standard log pattern
from log4j2 config, that is easy to read.