test: allow connecting to test DB via psql
https://www.testcontainers.org/modules/databases/jdbc does not set the
postgresql container username, password environment variables for us.
It also does not allow us to do so due to
https://github.com/testcontainers/testcontainers-java/issues/5028
Because of that the dhis role is not created. This is why we cannot
connect to the test DB containers via psql and dhis role/pw. Going back
to

https://www.testcontainers.org/test_framework_integration/manual_lifecycle_control/\#singleton-containers

allows us to connect to postgres again.