test: DataValueSetServiceTest leaves shared application context dirty (#11462)
* test: improve DataValueSetServiceIntegrationTest assertions

make sure to run all assertions on assertSuccessWithImportedUpdatedDeleted so we get as much context as we can
in case the import failed. Especially useful for flaky tests and tests only failing on GitHub. Add
the summary description in case the import was not successful.

summary.hasConflicts() checks the conflict count, so its redundant.

* test: DataValueSetServiceTest leaves shared application context dirty

DataValueSetServiceTest changes the BatchHandlerFactory to use a mock
without cleaning it up afterwards. This meant that if
DataValueSetServiceIntegrationTest ran after it no datavalues were
committed to the DB. Locally, DataValueSetServiceIntegrationTest ran
before DataValueSetServiceTest so tests did pass.

* test: enable DataValueSetServiceIntegrationTest