test: remove setDependency TECH-1194 (#11499)
* test: no need to set CurrentUserService

it is injected by Spring when needed

* test: use helpers in DataValueSetServiceTest

Same ones from DataValueSetServiceIntegrationTest since both tests can be merged at a
later point. Both are integration tests testing the DataValueSetService

* test: remove setDependency

EventPredictionService should be turned into a unit with mocks or integration test without mocks.
DataValueSetServiceTest can be merged with DataValueSetServiceIntegrationTest later on.

* test: remove unused

* test: migrate one integration to a unit test DataValueSetService

For the case that a DataValue is being updated with a DataValue with the
same value/comment. In this case we should not hit the DB. This can be
easily asserted with a Mock but not so in an integration test.

* test: assert expected ignored show up in report