ci: run all tests on Jenkins
in other versions the default profile which was activated by default
prevented integration tests from running on Jenkins.

This is not the cause in 2.36. Excluding the default profile using
'-default' prevents this from happening in 2.36 though.

Adapted pipelines to use the pom.xml instead of the pom-full.xml since having
2 has the risk of them diverging as they have. pom-full.xml only being used on
Jenkins and pom.xml locally and on GitHub.