test: ensure correct surefire version (#11588)
there are scenarios where we might run maven without any profile or a profile
that does not declare the surefire version in its build config. In that case maven
will resort to a default version which is rather old. This version does not
support JUnit 5 tests. So in case a JUnit 5 test would be added, it might not run.