test: be more lenient in cache header assertion (#11777)
A few tests like NO_CACHE and CACHE_1_MINUTE are enough to assert that
the CacheStrategy passed to configureResponse is making it into the max-age=60
response header value. Most other durations just delegate to TimeUnit
for making the calculation which we do not need to test.

CACHE_6AM_TOMORROW calls new Date(). So testing this we make 2 calls to
it once in the test for the expected value and once in the
implementation. Since we cannot control the time in the test, we can be
a bit more lenient in the expectation of the max-age seconds.