ci: Jenkins docker-compose does not support missing (#12089)
as pull_policy. Ideally we would add this to ensure the locally built (available) image
is used. We are using a tag that is not pushed to our Dockerhub repo. So this
is likely the behaviour anyway. It would just be safer and highlight the intent
of the e2e test setup better