ci: improve e2e test workflow [2.37] (#12098)
* chore: remove unused Dockerfile

* ci: improve e2e test workflow

allow running e2e tests using docker compose within one command. e2e tests
will be built automatically. Allow passing/pulling of an arbitrary image like
an image pushed from a PR. Such images will only be pulled if missing. This is
to avoid pulling an image in CI where we only want to test the images built on CI.

Improve the README by explaining all the ways in which the e2e tests can be run.
As well as how to potentially debug any failing e2e tests.

* ci: Jenkins docker-compose does not support missing

as pull_policy. Ideally we would add this to ensure the locally built (available) image
is used. We are using a tag that is not pushed to our Dockerhub repo. So this
is likely the behaviour anyway. It would just be safer and highlight the intent
of the e2e test setup better