ci(deps): ignore Spring framework 6 related updates
as Spring 6 requires minimum JDK 17 so it will require more effort/time until we update it.

For Spring and Spring security both need to be on the same major version
but are released independently.

Other libraries released by Spring follow a different release schedule
where the major version number differs.