fix: turn removeOnError into skipOnError
instead of iterating over hooks each iterating over entities and removing an invalid one from the bundle
iterate over entities then hooks and skip subsequent hooks if the hook deems the entity invalid and
has skipOnError set to true.

This is an intermediate step in a series of refactorings. The TrackerValidationHook interface will
not only contain default methods. First we need to move the orchestration from the AbstractTrackerDtoValidationHook
solely into the DefaultTrackerValidationService.