test: fix AtomicModeTest and ReportSummaryIntegrationTest
We previously did not find all errors due to the removeOnError logic.
This is why the incorrect test setup in ReportSummaryIntegrationTest was
hidden. We cannot add 2 active/completed enrollments into the particular
program.

AtomicModeTest now sees one more error which was previously not raised
as well due to the removeOnError logic. As
RelationshipsValidationHook.validateRelationshipConstraint now adds
error for the type constraint on the relationship not being met by the
TEI which has no trackedEntityType set