chore: move ValidationErrorReporter close to its usage
as its only used inside the validation packages.

This also creates a clearer separation to the report package. report is
the home for everything needed to create a TrackerImportReport of which
TrackerValidationReport is a part.

ValidationErrorReporter is only used to collect errors during validation
but not concerned with the representation (view) aspect.