fix: turn removeOnError into skipOnError DHIS2-14213 (#12457)
* fix: turn removeOnError into skipOnError

instead of iterating over hooks each iterating over entities and removing an invalid one from the bundle
iterate over entities then hooks and skip subsequent hooks if the hook deems the entity invalid and
has skipOnError set to true.

This is an intermediate step in a series of refactorings. The TrackerValidationHook interface will
not only contain default methods. First we need to move the orchestration from the AbstractTrackerDtoValidationHook
solely into the DefaultTrackerValidationService.

* chore: remove AbstractTrackerDtoValidationHook

as we moved its logic to DefaultTrackerValidationService and ValidationUtils

* test: fix AtomicModeTest and ReportSummaryIntegrationTest

We previously did not find all errors due to the removeOnError logic.
This is why the incorrect test setup in ReportSummaryIntegrationTest was
hidden. We cannot add 2 active/completed enrollments into the particular
program.

AtomicModeTest now sees one more error which was previously not raised
as well due to the removeOnError logic. As
RelationshipsValidationHook.validateRelationshipConstraint now adds
error for the type constraint on the relationship not being met by the
TEI which has no trackedEntityType set

* chore: remove empty method

* chore: remove default skipOnError and unused method

skipOnError is false by default

* chore: adapt javadoc of TrackerValidationHook

* test: prevent regression of removeOnError bug

at least when it comes to enrollment/event program check

* refactor: extract helpers from validate

* chore: remove todo