chore: remove unused duplicate reporting logic
I mistakenly felt like TrackerValidationReport and
ValidationErrorReporter should be merged. They have different responsibilities though.

TrackerValidationReport is for representational purposes only. It aggregates different reports of its
kind and presents them as JSON to the user. Its part of the TrackerImportReport.

ValidationErrorReporter responsibility is to collect errors/warnings that occur during the validation.

I think we need to find a better name