refactor: clean ValidationErrorReporter DHIS2-14213 (#12473)
* chore: remove modifiers which are added by lombok

* chore: collect invalid dtos in set

as we do not need to keep track of how many errors a dto has.

* chore: improve javadoc

* chore: track invalid dtos in enum map

as its more efficient since amount of keys is known beforehand

* chore: don't allow direct access to invalid dtos

its an internal structure the reporter keeps track of. Access is given via
its isInvalid methods.

* chore: group methods related to fields together

in call order. So if one like addErrorIf calls another one addErrorIf should be
above it. So we can read/navigate from top to bottom

* chore: name fields simply error and warning

* reportList makes it unclear that these are actually errors. warnings
  are also reports since both are TrackerError/WarningReports
* no need to add suffixes like list as this is clear from its type