fix: invalidation of tracker tree DHIS2-14213 (#12495)
collect valid entities (trackedEntities, enrollments, events, relationships) and errors explaining why other entities are invalid. The filter relies on preprocessing and the validation hooks that run before it doing the bulk of the work. It is only concerned with collecting

* entities which are valid i.e. have no validation errors found by previous hooks
* and entities which have valid parents (on CREATE/UPDATE) or valid children (DELETE)

All entities passing this filter should be persistable (CREATE/UPDATE/DELETE).