chore: remove unneeded prefix to TrackerValidationHook
Our classes should have a meaningful name that is as short as possible. There is a too short or
not meaningful. Here its clear that we are dealing with tracker as we are in the tracker module.
Its also clear that we are dealing with validation as we are in the validation package.

Using a noun that fits the interface methods is a common pattern. This leads to code
that can be written/read like prose. A validators validates ...