test: make assertHasError reusable using Validation
Now that we have the Validation interface implemented by errors and
warnings we can use it to create assertions on Validation. On top of
that we create small wrappers asserting for errors and warnings in
collection of Validations. On top of that we create wrappers asserting
on errors and warnings in the validation.Report and the
ValidationResult.

All use the same order of parameters with ValidationCode and if provided
a Validation.message.

Add assertHasWarning so its symmetric to asserting errors.