test: dont assert error count in unit tests (#12610)
Our validators are built to collect as many errors as possible. Unit tests are less brittle
if we only assert the error/warning our test case is concerned with is collected. If the
validator adds another error because we also trigger another case that is not an issue
in a unit test.

Our integration tests need to assert on the exact count since that is the final result
the user sees.