ci: log to STDOUT only in docker-compose setup (#12900)
as is standard in containerized environments. Logs can easily be
streamed from STDOUT. Logs can be redirected to files if needed.

Simplified the log layout to make them easier to read and parse.
Timestamps can be obtained by adding a flag to the docker CLI if needed.
The pattern allows streaming logs of a particular package. This was
previously achieved by logging into different files.