feat: use Converter to parse fields request parameter DHIS2-13731 (#12936)
* feat: use Converter to parse fields request parameter

This allows devs to access FieldPath in their controllers. Enables for example
only fetching things from the DB which will also be returned in the JSON after
field filtering has been applied. Currently, parsing needs to be done twice
as FieldPath are not exposed

* chore: allow passing parsed filters into FieldFilterService

so there is no need to parse twice

* chore: remove duplication in apply methods

* chore: add javadoc

* chore: no need for service id

as we don not implement an interface where we might need to use
a Qualifier