fix: tracker does not return fields in accordance with field filtering DHIS2-14783 (#13029)
* chore: order expressions in parameter declaration order

* chore: do not mutate the field paths

mutating the List<FieldPath> is surprising to the user of the FieldFilterService.
The parameter should be reusable by clients and considered immutable from the FieldFilterService perspective.

* feat: add FieldFilterService.filterIncludes

reusing the field filtering logic responsible to decide what fields are
included when filtering the JSON

* chore: remove children of exclusion in applyExclusions

* core: use filterIncludes in EventController

* core: use filterIncludes in EnrollmentController

* chore: use filterIncludes in TrackedEntity controller

* chore: review comments

* chore: inline method

its easier to read the getFieldPathMap call then to come up with a good name for this method

* chore: use parametrized test

* doc: add javadoc

* chore: use List.of

* docs: correct javadoc of FieldFilterParser

* chore: shorten names

* chore: remove todos