fix: tracker does not return fields in accordance with field filtering DHIS2-14783 [2.39] (#13076) (#13083)
* feat: use Converter to parse fields request parameter DHIS2-13731 (#12936)

* feat: use Converter to parse fields request parameter

This allows devs to access FieldPath in their controllers. Enables for example
only fetching things from the DB which will also be returned in the JSON after
field filtering has been applied. Currently, parsing needs to be done twice
as FieldPath are not exposed

* chore: allow passing parsed filters into FieldFilterService

so there is no need to parse twice

* chore: remove duplication in apply methods

* chore: add javadoc

* chore: no need for service id

as we don not implement an interface where we might need to use
a Qualifier

* fix: Remove list wrapper in field filter service [DHIS2-14743]

* fix: tracker does not return fields in accordance with field filtering DHIS2-14783 (#13029)

* chore: order expressions in parameter declaration order

* chore: do not mutate the field paths

mutating the List<FieldPath> is surprising to the user of the FieldFilterService.
The parameter should be reusable by clients and considered immutable from the FieldFilterService perspective.

* feat: add FieldFilterService.filterIncludes

reusing the field filtering logic responsible to decide what fields are
included when filtering the JSON

* chore: remove children of exclusion in applyExclusions

* core: use filterIncludes in EventController

* core: use filterIncludes in EnrollmentController

* chore: use filterIncludes in TrackedEntity controller

* chore: review comments

* chore: inline method

its easier to read the getFieldPathMap call then to come up with a good name for this method

* chore: use parametrized test

* doc: add javadoc

* chore: use List.of

* docs: correct javadoc of FieldFilterParser

* chore: shorten names

* chore: remove todos

* chore: remove duplication

to prevent any future drift of writing to a list of ObjectNodes
and the JSON stream

---------

Co-authored-by: Lars Helge Øverland <lars@dhis2.org>