chore: remove unused href field from new tracker event export (#13088)
it was added when /tracker/event was added in

https://github.com/dhis2/dhis2-core/pull/7103

According to Giuseppe he only added it because it was part of the old
tracker event endpoint for GET. No other GET endpoint in old and new
tracker exposes this field.

'href' was returned on GET if fields query parameter contained value
'href' or '*'. In new tracker 'href' is only returned on GET
/tracker/events until 2.36 after which it did not work anymore. So far
no one has reported an issue.