fix: Options for run_api bash script for embedded Jetty (#13188)
Refactors the run_api.sh script. Moves code into functions. Adds support for options with values, instead of relying on the order of arguments on the command line. Adds an option for setting the DHIS2_HOME variable. Adds an option for skipping compilation, to avoid having to press a key on every run.