chore: remove unused code
non-super users will not get an event they do not have access to
i.e. an event with a category option they do not have access to due to
query added by getCategoryOptionComboQuery. Checking the number of COs
from getAttributeCategoryOptions() matches the getOptionSize might
have been an attempt to address ACL. With the updated query
the COs and option size come from the same aggregation.