chore: make new tracker relationship web independent of dhis-service-dxf2 TECH-1515 (#13283)
* chore: make new tracker relationship web independent of dxf2

dxf2 services are built on top of dhis-api services. So we cannot simply
rely on dhis-api services as dxf2 services do some ACL checks and other
types of mappings. Duplicate the retrieve part of the dxf2 relationship
service. The new tracker relationship services still depends on dxf2 as
it needs dxf2 services to fetch details on the nested entities like
trackedEntity, enrollment and event. We can remove the new tracker
relationship service dependency on dxf2 once we have created new tracker
services for those entities.

In the meantime we have to go from dxf2 to dhis-api in our service and
then from dhis-api to tracker view in the controller.

* chore: use RelationshipStore directly in DefaultRelationshipService

as the dhis-api RelationshipService is just a thin wrapper around the store.
Interestingly skipAccessValidation was always false and even discarded in the
dhis-api RelationshipStore