feat: Add displayName to WebLocale and use user-selected language [DHIS2-14907] (#13874)
* feat: Update the ui and db locales endpoints [DHIS2-14907]

This update returns the WebLocale objects with a new structure and makes sure it's displayed in the user's selected language and not the server-selected language. Previously, the WebLocale object only had 2 properties, 'locale' and 'name'. 'locale' is the locale e.g 'en', 'fr' etc. 'name' was the display name for that locale which was always set to the server-selected language. 'name' is now used to display the language in its own language e.g. English, français, espanol etc. 'displayName' is the newly-added property which will display the locale language in the user-selected locale.

* feat: Add tests for db locales endpoint

Added tests for db locales endpoint and added JsonWebLocale interface to the dhis-support-test module

* feat: Add tests to confirm correct language precedence

---------

Co-authored-by: David Mackessy <davidmackessy@Davids-MacBook-Pro.local>