chore: move old tracker dxf2 code into deprecated package (#13901)
so that one day we can look for 'deprecated.tracker' and remove old tracker
by deleting a couple of directories. We chose 'deprecated.tracker' instead
of 'tracker.deprecated' as we wanted to increase the distance between
old and new tracker. In some modules like dhis-web-api old and new tracker
live together. We want to keep old tracker away from new. As it was and still
is a considerable effort to detach the two.