feat: Add copy endpoint and service to copy a Program
Added a copy endpoint in the ProgramController to allow copying of a Program by its uid. Added CopyService which gets a copy of the Program. Added copyOf methods in Program, ProgramStage, ProgramStageSection and ProgramStageDataElement which return copies of each instance, with shallow and deep copies of proerties where appropriate. Updated some hibernate properties to allow cascade saving where necessary.