wip: should we do this separately?
the issue with category options is that we even use string concatenated
options using ';' in SQL queries. I hope we are not storing this string
in the DB. We also return it in the event.attributeCategoryOptions using
';'. So the question is should we then do this all at once. This would
also be a breaking change unless we put in the effort of returning a
different payload based on versioning.