chore: document deprecated /tracker/relationships?tei parameter
Our OpenAPI generation does not provide the same mechanism as Spring
does to have one field 'trackedEntity' with 2 setters providing the old
and new query parameter. We therefore need to add the deprecated field
so we can document it in our spec. We will be able to remove it in 2.42.