fix: /tracker/trackedEntities?order=createdAt (#14631)
was ordering by tracked entity id instead of created column.

Lowercase column names in OrderColumn mapper.
Fix tests which asserted using the wrong order or assumed we can set the lastupdated date. lastupdated
is automatically set by HibernateGenericStore. So to assert we can order by that column we can only
update the entity in some order and expect that to be returned