chore: remove grid remnant from new tracker events store (#14756)
When we copied the JdbcEventStore to new tracker we removed the
getEventsGrid() as this is only used in old tracker /events/query.
We forgot to remove buildGridSql(params, mapSqlParameterSource) which is
used in getting the event count in case there are filters. We should
remove this logic as the count should be applied on exactly the same
query as the one returning the actual events. getEventSelectQuery is
used in buildSql.