test: add order tests for tracked entities
and move order and pagination tests into one place. Our entities are
related TE -> EN -> EV. Our endpoints allow ordering of a parent or
child to affect the outer order. Testing thus becomes easier and clearer
if we test ordering and pagination in the same test class. We can see
for example why TE are ordered the way they are if ordered by enrolledAt
date as we will have a test for
/tracker/trackedEntities?order=enrolledAt and one for
/tracker/enrollment?order=enrolledAt.