chore: move handleSortAttributes into TrackedEntityQueryParams  TECH-1611 (#14837)
* test: /tracker/trackedEntities?filter and attribute

* chore: move handleSortAttributes into TrackedEntityQueryParams

similar to EventSearchParams. It should be hard to misuse TrackedEntityQueryParams. Users
of it should not have to remember to add the order tea into attributes with
all the conditions of when to do it

* chore: simplify tracked entities store query generation

set the alias in the subquery as we do in event. Reuse that in the orderable fields
value

* chore: sort tracked entities desc by default

like events

* chore: use simpler alias