chore: do not allow disabling Java formatter (#14844)
The code formatting is not consistent when formatting directly in an IDE
(Eclipse/Intellij) using the xml formatter config then compared with
formatting via the Spotless plugin. The formatter xml configs that are
provided even in the google formatter project are not reliable. Seems
like the suggestion is that they should actually be removed from the
google java formatter project. The google formatter cannot be turned of
by design. Removing the ability to turn it off via spotless allows us to
use the google java formatter plugin in Intellij.