fix: Use set for category to category combo association (#15186)
Having many-to-many bi-directional associations with an ordered collection List on both sides does seem to be meaningless from the inverse side. From what I can see, supporting ordering on both sides using a single link table and sort order column can not be done in a meaningful way. So it would be better to change the Java collection type on the inverse side to an unordered Set. This will remove the conflicting ordering needs.