chore: move handling of total count and pagination to store
having the logic of fetching + 1 pages in the store and the removal of
the extra page/elements in the service is error prone. This should be
handled in the store alone and transparent to any service.

Total count is expensive and should thus be discouraged. Do not expose
this as part of the service as its not needed right now.