chore: move handling of total count and pagination to store TECH-1638 (#15223)
* chore: remove unused parameter

* chore: move handling of total count and pagination to store

having the logic of fetching + 1 pages in the store and the removal of
the extra page/elements in the service is error prone. This should be
handled in the store alone and transparent to any service.

Total count is expensive and should thus be discouraged. Do not expose
this as part of the service as its not needed right now.

* chore: fix linting issue

* chore: incorporate PR feedback

* chore: do not expose getting total count from event store

as its only used internally to return the count in the page