fix: validation of /tracker/trackedEntities?order TECH-1648 [2.38] (#16114)
* fix: validation of /tracker/trackedEntities?order=

In https://dhis2.atlassian.net/browse/TECH-1611 we harmonized ordering
trackedEntities in 2.41. This now allows ordering only by field names of
new tracker. Previously, users could also order by 'created' since
TrackedEntityInstanceQueryParams.OrderColumn is responsible for mapping
field names and is used by new and old tracker.

This change rejects any order request for non supported fields with a 400 like in 2.41.

Note that this change differs to the one on 2.40 as the same
TrackedEntityCriteriaMapper class is used by old and new tracker in
2.39. Validation is thus done in the controller so we do not affect old
tracker.

* fix: exclude raw order from mapper

otherwise the order fields are duplicated