fix: reject invalid order parameter values DHIS2-16935 (#16673)
* fix: reject invalid order parameter values

Rename OrderCriteria factory to valueOf so its used by Springs conversion
system. The OrderCriteriaParamEditor is not needed. In fact its even causing the
issue that "order=createdAt:desc,event:asc" is not split by Spring and passed
as is to the PropertyEditor. Whatever mechansim is used to convert Spring
suggests throwing an IllegalArgumentException in case conversion failed. We returned
null which lead to successful conversions with null elements

* chore: reject empty order fields

* test: asssert on actual types and their order

instead of strings. Add tests for passing comma separated values in one
order param. Add test for the mixed case of comma separated values in one
order param with a repeated param

* test: TE workinglist separates order using ; instead of ,

* chore: UID does not need a PropertyEditor

as it has a contructor from String -> UID its automatically
picked up by Spring

* chore: cannot order by orgUnit using this implementation

we need to update this to the new tracker mechanism. Not sure if
these workinglists are also used by old tracker though

* test: UID binding

* test: fix

* test: clean